Refactor: move logging support into dedicated module
